refactor: move contact groups and distribution-lists in their own folders under views/
- move components together
