+++ src/views/contact-groups/GroupsAppView.tsx
@@ -11,8 +11,8 @@ import { trimEnd } from 'lodash';
 import { Redirect, Route, Switch, useRouteMatch } from 'react-router-dom';
 
 import CGView from './cg-view';
-import { DistributionListsView } from './distribution-lists-view';
-import { RouteParams, ROUTES, ROUTES_INTERNAL_PARAMS } from '../constants';
+import { RouteParams, ROUTES, ROUTES_INTERNAL_PARAMS } from '../../constants';
+import { DistributionListsView } from '../distribution-list/distribution-lists-view';
 
 const AppView = (): React.JSX.Element => {
 	const { path, params, url } = useRouteMatch<RouteParams>();
