+++ src/views/distribution-list/SecondaryBarView.test.tsx
@@ -9,8 +9,8 @@ import React from 'react';
 import 'jest-styled-components';
 
 import SecondaryBarView from './SecondaryBarView';
-import { screen, setupTest, within } from '../carbonio-ui-commons/test/test-setup';
-import { TESTID_SELECTORS } from '../constants/tests';
+import { screen, setupTest, within } from '../../carbonio-ui-commons/test/test-setup';
+import { TESTID_SELECTORS } from '../../constants/tests';
 
 // function spyUseNavigation(navigateTo: jest.Mock): void {
 // 	jest.spyOn(useNavigationHook, 'useNavigation').mockReturnValue({
